[Feat] 레이저 장애물 게임 오브젝트 구성
