+++ Katana_ZERO/Assets/Scripts/Character/Enemy/Boss/Kissyface/States/KissyLunge.cs
@@ -1,5 +1,5 @@
 using UnityEngine;
-using LiteralRepository;
+using HelperFuncRepository;
 using static KissyfaceAnimInvoker;
 
 public class KissyLunge : BossStateMachine
@@ -24,7 +24,7 @@ public class KissyLunge : BossStateMachine
     {
         base.OnStateUpdate(animator, stateInfo, layerIndex);
 
-        moveVec = SecondBezier( startPos, _controlPos, playerPos, stateInfo.normalizedTime );
+        moveVec = HelperFunc.SecondBezier( startPos, _controlPos, playerPos, stateInfo.normalizedTime );
 
         rigid.MovePosition( moveVec );
     }
@@ -32,16 +32,5 @@ public class KissyLunge : BossStateMachine
     override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateExit(animator, stateInfo, layerIndex);
-    }
-
-    private Vector2 FirstBezier( Vector2 p0, Vector2 p1, float t )
-    {
-        return Vector3.Lerp( p0, p1, t );
-    }
-    private Vector2 SecondBezier( Vector2 p0, Vector2 p1, Vector2 p2, float t )
-    {
-        Vector2 m0 = FirstBezier( p0, p1, t );
-        Vector2 m1 = FirstBezier( p1, p2, t );
-        return FirstBezier( m0, m1, t );
     }
 }
